hook wp_handle_upload instead
of wp_handle_upload_prefilter. This is because of the Enable Media Replace plugin, which itself hooks wp_handle_upload, but not the prefilter. Other plugins may do the same, so making this change may improve our compatiibility with those plugins as well.  The primary benefit of the prefilter is that it allows you to reject the request based on custom criteria before the upload happens. We don't actually need to do that here, so the extra compatibility may be a more important benefit.