Prevents accidental S3 deletion on site delete
Adds a check to ensure that when a site is deleted, the S3 assets are only deleted if the site exists within the current network. This prevents accidental data loss if a site has been incorrectly copied between environments and references to the original site remain. Also logs attempted deletions of sites not in the current network for auditing.